Rename primitives to types
I'm gonna put factories in there for different CRDT types, so the name
primitives doesn't fit anymore.
